@@ -1,3 +1,3 @@
-module.exports = {
+export default {
   '*.{js,jsx,vue,css,scss}': 'npm run format-lint',
 }
